requirements.txt
# Elderly Care Smart Home System Requirements

# Core GUI Framework
PyQt5>=5.15.0

# MQTT Communication
paho-mqtt>=1.6.0

# Data Processing and Analysis
pandas>=1.3.0
numpy>=1.21.0

# Plotting and Visualization
matplotlib>=3.5.0
pyqtgraph>=0.12.0

# Debugging and Logging
icecream>=2.1.0

